Organizing and cleaning code
