+++ Crypt Creeper/Crypt Creeper.xcodeproj/xcuserdata/dmoiz.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,8 +14,8 @@
             filePath = "Crypt Creeper/Presentation/SignUp/Factions/FactionView-ViewModel.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "80"
-            endingLineNumber = "80"
+            startingLineNumber = "45"
+            endingLineNumber = "45"
             landmarkName = "signUp(user:completion:)"
             landmarkType = "7">
          </BreakpointContent>
