+++ Crypt Creeper/Crypt Creeper/Presentation/Login/LoginView-ViewModel.swift
@@ -23,46 +23,6 @@ extension LoginView {
         let userDefaults = UserDefaults.standard
         private var token: String = ""
         
-        struct User: Codable {
-            var name: String
-            var password: String
-        }
-        
-        enum NetworkError: Error {
-            case invalidUrl
-            case invalidResponse
-            case invalidData
-            case requestFailed
-            case authenticationError
-            case badRequest
-            case outdated
-            case failed
-            case unknown
-            
-            var localizedDescription: String {
-                switch self {
-                case .invalidUrl:
-                    return "The URL is not valid"
-                case .invalidResponse:
-                    return "The response received from the server is not valid"
-                case .invalidData:
-                    return "The data received from the server is not valid"
-                case .requestFailed:
-                    return "The request failed"
-                case .authenticationError:
-                    return "There was an authentication error"
-                case .badRequest:
-                    return "The request is not valid"
-                case .outdated:
-                    return "The request is outdated"
-                case .failed:
-                    return "The request failed for an unknown reason"
-                case .unknown:
-                    return "An unknown error occurred"
-                }
-            }
-        }
-        
         func login(completion: @escaping (Result<String, Error>) -> Void) {
             let url = "http://127.0.0.1:8000/api/user/login"
             let dictionary: [String: Any] = [
@@ -78,7 +38,7 @@ extension LoginView {
                 
                 // Verificar si se recibió una respuesta válida
                 guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                    completion(.failure(NetworkError.invalidResponse))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                     return
                 }
                 
@@ -88,10 +48,10 @@ extension LoginView {
                     if let responseJson = json as? [String: Any], let message = responseJson["Token"] as? String {
                        completion(.success(message))
                    } else {
-                       completion(.failure(NetworkError.invalidData))
+                       completion(.failure(NetworkError.networkErrorEnum.invalidData))
                    }
                } catch {
-                   completion(.failure(NetworkError.invalidData))
+                   completion(.failure(NetworkError.networkErrorEnum.invalidData))
                }
             }
         }
