+++ Crypt Creeper/Crypt Creeper/Presentation/SignUp/Factions/FactionView-ViewModel.swift
@@ -18,41 +18,6 @@ extension FactionsView {
         let userDefaults = UserDefaults.standard
         private var token: String = ""
         
-        enum NetworkError: Error {
-            case invalidUrl
-            case invalidResponse
-            case invalidData
-            case requestFailed
-            case authenticationError
-            case badRequest
-            case outdated
-            case failed
-            case unknown
-            
-            var localizedDescription: String {
-                switch self {
-                case .invalidUrl:
-                    return "The URL is not valid"
-                case .invalidResponse:
-                    return "The response received from the server is not valid"
-                case .invalidData:
-                    return "The data received from the server is not valid"
-                case .requestFailed:
-                    return "The request failed"
-                case .authenticationError:
-                    return "There was an authentication error"
-                case .badRequest:
-                    return "The request is not valid"
-                case .outdated:
-                    return "The request is outdated"
-                case .failed:
-                    return "The request failed for an unknown reason"
-                case .unknown:
-                    return "An unknown error occurred"
-                }
-            }
-        }
-        
         func signUp(user: UserModel, completion: @escaping (Result<String, Error>) -> Void) {
             let url = "http://127.0.0.1:8000/api/user/create"
             let dictionary: [String: Any] = [
@@ -71,7 +36,7 @@ extension FactionsView {
                 
                 // Verificar si se recibió una respuesta válida
                 guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                    completion(.failure(NetworkError.invalidResponse))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                     return
                 }
                 
@@ -81,10 +46,10 @@ extension FactionsView {
                     if let responseJson = json as? [String: Any], let message = responseJson["Token"] as? String {
                        completion(.success(message))
                    } else {
-                       completion(.failure(NetworkError.invalidData))
+                       completion(.failure(NetworkError.networkErrorEnum.invalidData))
                    }
                } catch {
-                   completion(.failure(NetworkError.invalidData))
+                   completion(.failure(NetworkError.networkErrorEnum.invalidData))
                }
             }
         }
