+++ Crypt Creeper/Crypt Creeper/Presentation/SignUp/Factions/FactionsView.swift
@@ -7,30 +7,6 @@
 
 import SwiftUI
 
-struct UserModel {
-    let username: String
-    let email: String
-    let password: String
-    var factionId: Int
-    let profilePic: UIImage
-    
-    init() {
-        self.username = ""
-        self.email = ""
-        self.password = ""
-        self.factionId = 1
-        self.profilePic = UIImage()
-    }
-    
-    init(username: String, email: String, password: String, factionId: Int, profilePic: UIImage) {
-        self.username = username
-        self.email = email
-        self.password = password
-        self.factionId = factionId
-        self.profilePic = profilePic
-    }
-}
-
 struct FactionsView: View {
     
     @State var userModel: UserModel = .init()
