+++ Crypt Creeper/Crypt Creeper/UserSettings/SettigsViewModel.swift
@@ -8,50 +8,6 @@
 import Foundation
 import UIKit
 
-struct user : Codable {
-    let name : String
-    var password : String
-    var email : String
-    var faction_id : Int
-    var profile_pic : String
-}
-enum NetworkError: Error {
-    case invalidUrl
-    case invalidResponse
-    case invalidData
-    case requestFailed
-    case authenticationError
-    case badRequest
-    case outdated
-    case failed
-    case unknown
-    
-    var localizedDescription: String {
-        switch self {
-        case .invalidUrl:
-            return "The URL is not valid"
-        case .invalidResponse:
-            return "The response received from the server is not valid"
-        case .invalidData:
-            return "The data received from the server is not valid"
-        case .requestFailed:
-            return "The request failed"
-        case .authenticationError:
-            return "There was an authentication error"
-        case .badRequest:
-            return "The request is not valid"
-        case .outdated:
-            return "The request is outdated"
-        case .failed:
-            return "The request failed for an unknown reason"
-        case .unknown:
-            return "An unknown error occurred"
-        }
-    }
-}
-
-
-
 class SettingsViewModel : ObservableObject {
     
     @Published var userName = ""
@@ -88,7 +44,7 @@ class SettingsViewModel : ObservableObject {
             }
             // Verificar si se recibió una respuesta válida
             guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                completion(.failure(NetworkError.invalidResponse))
+                completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                 return
             }
             
@@ -98,10 +54,10 @@ class SettingsViewModel : ObservableObject {
                 if let responseJson = json as? [String: Any], let message = responseJson["message"] as? String {
                     completion(.success(message))
                 } else {
-                    completion(.failure(NetworkError.invalidData))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidData))
                 }
             } catch {
-                completion(.failure(NetworkError.invalidData))
+                completion(.failure(NetworkError.networkErrorEnum.invalidData))
             }
         }
     }
@@ -122,7 +78,7 @@ class SettingsViewModel : ObservableObject {
             }
             // Verificar si se recibió una respuesta válida
             guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                completion(.failure(NetworkError.invalidResponse))
+                completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                 return
             }
             
@@ -132,10 +88,10 @@ class SettingsViewModel : ObservableObject {
                 if let responseJson = json as? [String: Any], let message = responseJson["message"] as? String {
                     completion(.success(message))
                 } else {
-                    completion(.failure(NetworkError.invalidData))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidData))
                 }
             } catch {
-                completion(.failure(NetworkError.invalidData))
+                completion(.failure(NetworkError.networkErrorEnum.invalidData))
             }
         }
     }
@@ -156,7 +112,7 @@ class SettingsViewModel : ObservableObject {
             }
             // Verificar si se recibió una respuesta válida
             guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                completion(.failure(NetworkError.invalidResponse))
+                completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                 return
             }
             // Procesar la respuesta
@@ -165,10 +121,10 @@ class SettingsViewModel : ObservableObject {
                 if let responseJson = json as? [String: Any], let message = responseJson["message"] as? String {
                     completion(.success(message))
                 } else {
-                    completion(.failure(NetworkError.invalidData))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidData))
                 }
             } catch {
-                completion(.failure(NetworkError.invalidData))
+                completion(.failure(NetworkError.networkErrorEnum.invalidData))
             }
         }
     }
@@ -182,7 +138,7 @@ class SettingsViewModel : ObservableObject {
         // Verificar si se recibió una respuesta válida
         NetworkHelper.shared.requestProvider(url: url) { data, response, error in
             guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                completion(.failure(NetworkError.invalidResponse))
+                completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                 return
             }
     
@@ -192,10 +148,10 @@ class SettingsViewModel : ObservableObject {
                 if let responseJson = json as? [String: Any], let message = responseJson["message"] as? String {
                     completion(.success(message))
                 } else {
-                    completion(.failure(NetworkError.invalidData))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidData))
                 }
             } catch {
-                completion(.failure(NetworkError.invalidData))
+                completion(.failure(NetworkError.networkErrorEnum.invalidData))
             }
         }
         
@@ -215,7 +171,7 @@ class SettingsViewModel : ObservableObject {
             }
             // Verificar si se recibió una respuesta válida
             guard let data = data, let httpResponse = response as? HTTPURLResponse, (200..<599).contains(httpResponse.statusCode) else {
-                completion(.failure(NetworkError.invalidResponse))
+                completion(.failure(NetworkError.networkErrorEnum.invalidResponse))
                 return
             }
             
@@ -225,10 +181,10 @@ class SettingsViewModel : ObservableObject {
                 if let responseJson = json as? [String: Any], let message = responseJson["message"] as? String {
                     completion(.success(message))
                 } else {
-                    completion(.failure(NetworkError.invalidData))
+                    completion(.failure(NetworkError.networkErrorEnum.invalidData))
                 }
             } catch {
-                completion(.failure(NetworkError.invalidData))
+                completion(.failure(NetworkError.networkErrorEnum.invalidData))
             }
         }
     }
